Voting
An election is contested by five candidates. The candidates are numbered 1 to 5 and the voting is done by marking the candidate number on the ballot papers. Write a program to read the ballots and count the votes cast for each candidate member variable count. In case, a number read is outside the range of 1 to 5, that ballot should be considered as a 'spoilt ballot', and the program should also count the number of spoilt ballots.