Operator Overloading
Create a class Distance, which accepts data in feet and inches adds two distances, and displays the members of the distance object in the appropriate form. Test the class in the main program by creating objects function Overloading in C++
You can have multiple definitions for the same function name in the same scope. The definition of the function must differ from each other by the types and/or the number of arguments in the argument list. You cannot overload function declarations that differ only by return type. d1 and d2 of type distance, accept data for each object and add them then display them. (Operator Overloading)

Function Overloading in C++
You can have multiple definitions for the same function name in the same scope. The definition of the function must differ from each other by the types and/or the number of arguments in the argument list. You cannot overload function declarations that differ only by return type.