Friend Functions
 Write object-oriented C++ program Create two classes Minutes that stores time in minutes and Seconds that stores time in seconds. Read values for the class objects and add one object of Minutes with another object of Seconds. Use friend function to carry out the addition operation and display the resultant time in hours, minutes and seconds.

A friend function is a function that is specified outside a class but has the ability to access the class members’ protected and private data. A friend can be a member’s function, function template, or function, or a class or class template, in which case the entire class and all of its members are friends.