Multilevel inheritance
Write object-oriented C++ program to define a base class Student with data member name. Intermediate Derived class Details with data members telephone no. Another Derived class Stats with data member height. Program should display details of a student(Multilevel inheritance)

Inheritance is a process in which one object acquires all the properties and behaviors of its parent object automatically. In such way, you can reuse, extend or modify the attributes and behaviors which are defined in other class.
In C++, the class which inherits the members of another class is called derived class and the class whose members are inherited is called base class. The derived class is the specialized class for the base class.