Virtual Function
Create a base class in C++ called shape. Use this class to store two double type values that could be used to compute the area of figures. Derive two specific classes called triangle and rectangle from the base shape. Add to the base class a member function get_data () to initialize base class data members and another member function display_area () to compute and display the area of figures. Make display_area () a virtual function and redefine this function in the derived classes to suit their requirements. Using these classes, design a program that will accept the dimensions of a triangle or a rectangle interactively and display the area.